backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
coverage==7.6.3
cron-descriptor==1.4.5
cryptography==43.0.3
debugpy==1.8.7
defusedxml==0.8.0rc2
distlib==0.3.9
Django==4.2.13
django-bootstrap-v5==1.0.11
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-filter==24.3
django-rest-knox==5.0.2
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docker==7.1.0
exceptiongroup==1.2.1
filelock==3.16.1
flower==2.0.1
gunicorn==23.0.0
h11==0.14.0
humanize==4.9.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pillow==10.3.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.1.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.3
pytest-django==4.9.0
pytest-docker-tools==3.1.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
pywin32==308
redis==5.0.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
selenium==4.21.0
selenium-stealth==1.0.6
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stealth==0.0.1
tenacity==9.0.0
tornado==6.4.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.6
wcwidth==0.2.13
websockets==13.0.1
wsproto==1.2.0
